Seperated state logic in dedicated hook
